Format and add some additional values to getLocationInfo method
Made a couple useful data like shortname and urlname.  Add all the
amenity fields and values.  Find all the applicable amenities and add
to object as array.

Not sure if I need to include the whole amenity list, but it was
already in the object and didn't feel need to remove.

QUESTION: Is this a good way to pass this additional data around - ie.
the array?